refactor(audience): make TimedBossBar a concrete class
Drop the interface + TimedBossBarImpl split — one implementation does
not earn a parallel type, and the factory keeps construction internal.
